CI: Run package/check on all packages
This will check some basic properties on all package Makefiles before
the full build. It checks for example if the PKG_HASH and
PKG_MIRROR_HASH are correct and will fail if it is not correct. This
should detect such problems before they get merged.

The check step will use its own configuration independent if we use an
external or internal toolchain. We have to reset it back in the build
process.

This will always be executed when we do a full package build, the check
is pretty fast.

Signed-off-by: Hauke Mehrtens <hauke@hauke-m.de>