CI: reusable_build: outputs ccache_tag name
Also outputs ccache_tag name to permit user of this reusable workflow to
derive the ccache name for upload.

Outputs from reusable workflow is possible and supported by have a
limitation with matrix strategy. With matrix strategy the outputs are
taken only by the latest matrix run. This is not a problem as ccache tag
doesn't change and is always the same. (as it's based on github
variables based on base branch)

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>