CI: issue-labeller: always use updated version of dump target info
Always use updated version of dump target info on validating the device.
There is currently a problem where the script might be not avaliable or
outdated as we checkout reported release version in the BUG issue to
make sure the device actually exist for the reported release.

To solve this, we always checkout the latest version for the branch of
the release reported. It's still required to backport any update to the
script. (this is done to force keeping the script in sync across stable
branch and don't let it rot)

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>