kernel: fix change detection for kernel builds
Targets were previously added for testing if one of the following three
conditionn is true regarding the changed files:

- changes contain anything but `target/linux/.*/`
- changes in `target/linux/generic`
- changes contain `$TARGET`

The first condition is true for nearly any change, causing Kernels being
tested all the time. The new logic changes the conditions to one of the
following two:

- changes in `target/linux/generic`
- changes `target/linux/$TARGET`

In a quick test, this reduced tested kernel from $ALL to literally just
the one which was changed.

Signed-off-by: Paul Spooren <mail@aparcar.org>