CI: kernel: match whole-word for target kernels detection
The is currently a flwa in how the target kernel version are detected
for the changes.

For the corner case of 2 subtarget with one the prefix of the other:
- lantiq/xway
- lantiq/xway_legacy

Using a simple grep makes the kernel version for the target to count
both subtarget instead of only the specific one.

This caused the generation of a broken JSON and failing of the CI matrix
creation.

To fix this use the -w option of grep to match the whole word and
prevent these kind of corner case.

Fixes: 99d722a7fe03 ("CI: kernel: improve target/subtarget selection logic")
Fixes: b29e8c690352 ("CI: kernel: add support for testing testing kernel version if affected")
Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>