CI: kernel: improve affected kernel version detection logic
The current parsing to get the affetcted kernel version from the changed
files outputs multiple entry of the same kernel version (Example "6.1 6.6 6.1
6.6 6.1")

This is caused by uniq not removing the duplicate entry if they are not
one after another.

To fix this, sort the list before calling uniq to make the tool correctly
remove all the duplicate entry.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>