CI: multi-arch-test-build: fix error in detecting changed packages
Fix error in detecting changed packages if the packge are present in the
root directory instead of a subdirectory.

This is the case for routing feeds that have packages directly in the
root directory.

This caused a problem in detecting the packages as the sed regex didn't
account for this.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>