multi-arch-test-build: add generic package tests
Check if any of the package executables return the expected version when
called with a generic list of flags, e.g. --version, -v, etc.

Check if executable and library symlinks are valid.

Check if executables are marked as such.

Check for hardcoded paths.

Check if binaries are stripped.

Check if all shared linked libraries are installed.

Check if libraries have sonames and if so, if they have a soname
symlink.

Generic tests are both enabled and forced by default, i.e. they will run
even if a package-specific test is present.

Signed-off-by: George Sapkin <george@sapk.in>